vpn
The initial motivation for writing this tiny program was to create a
command-line utility I could call from my xmobarrc, since adding a
plugin to xmobar is actually pretty cumbersome.

All this currently does is check whether a vpn client is running, and if
so, print out a padlock icon (in FontAwesome).